pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rby1-workbench"
version = "0.1.0"
description = "RB-Y1 workbench for visualization, calibration, and control workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3",
    "numpy>=1.24",
    "omegaconf>=2.3",
    "rerun-sdk>=0.20",
    "rby1-sdk",
]

[project.scripts]
rby1-visualize-robot = "rby1_workbench.apps.visualize_robot:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
rby1_workbench = ["conf/*.yaml"]
